requirements.txt
pandas
openpyxl
numpy
scikit-learn
transformers
torch
sentence-transformers
faiss-cpu
matplotlib
seaborn
tqdm